refactor(app): repack frame_size to (i16, i16) to hold App at 34,944 B
Grouping fields into components introduced 4 B of intra-component tail
padding (RideEngine +1, HostPending +3), pushing size_of::<App>() from
34,944 to 34,952 on thumbv8m (align-8 rounding). Narrowing the recorded
panel size to i16 per dimension — a few hundred pixels, bounded far below
i16::MAX (the #810 u16-repack precedent) — reclaims 4 real bytes and lands
App back at exactly 34,944 B; the default board profile's resident set and
allocation report match the committed baseline byte-for-byte again.

+++ firmware/obc-app/src/app.rs
@@ -2126,7 +2126,7 @@ impl App {
     {
         // Record the panel size for the screen ticks' region reporting (`advance_animations`) —
         // the one place every host states its real frame dimensions.
-        self.ui.frame_size = (w as i32, h as i32);
+        self.ui.frame_size = (w as i16, h as i16);
         // Drain the one-shot region clip (see `set_render_clip`) — `None` on every normal frame.
         let render_clip = self.ui.render_clip.take();
